Split implementation into logically split files.
