+++ docx_to_pdf
@@ -1,4 +1,4 @@
 #!/bin/bash
 SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]:-$0}"; )" &> /dev/null && pwd 2> /dev/null; )";
 
-python3 "$SCRIPT_DIR"/doc_to_pdf.py "$@"
+python3 "$SCRIPT_DIR"/src/doc_to_pdf.py "$@"
